backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL driver

# Redis caching (hiredis for performance)
redis==5.0.1
hiredis==2.3.2

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0

# HTML/PDF parsing
beautifulsoup4==4.12.3
lxml==5.1.0
pdfplumber==0.10.3
pytesseract==0.3.10

# Rate limiting
aiolimiter==1.1.0
slowapi==0.1.9

# Optimization
python-constraint==1.4.0

# Retry and circuit breaker
tenacity==8.2.3
pybreaker==1.0.1

# Structured logging
structlog==24.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2

# Monitoring and metrics
psutil==5.9.8
prometheus-client==0.20.0
